Move Role to dedicated file
Relates to #118
